refactor: Move to data dir
