+++ lib/features/session/data/session.freezed.dart
